pyproject.toml
[tool.poetry]
name = "raddar"
version = "0.1.0"
description = ""
authors = ["Tioborto"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.61.1"
uvicorn = "^0.12.2"
detect-secrets = "^0.14.3"
async-exit-stack = "^1.0.1"
async_generator = "^1.10"
SQLAlchemy = "^1.3.20"
GitPython = "^3.1.9"
pydantic = {extras = ["dotenv"], version = "^1.6.1"}

[tool.poetry.dev-dependencies]
black = "^20.8b1"
isort = "^5.6.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
